backend/requirements.txt
absl-py==2.0.0
alembic==1.12.0
asttokens==2.4.0
azure-core==1.29.4
azure-identity==1.14.0
azure-storage-blob==12.18.2
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
cbor==1.0.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chest==0.2.3
click==8.1.7
cloudpickle==2.2.1
comm==0.1.4
contourpy==1.1.1
cryptography==41.0.4
cwl-eval==1.0.12
cycler==0.12.1
databricks-cli==0.18.0
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
docker==6.1.3
docopt==0.6.2
dpu-utils==0.6.1
elastic-transport==8.4.1
elasticsearch==8.10.0
entrypoints==0.4
executing==2.0.0
faiss-gpu==1.6.5
filelock==3.12.4
Flask==2.3.3
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
fonttools==4.43.1
fsspec==2023.9.2
function_parser @ git+https://github.com/g-lopes/function_parser
future==0.18.3
gensim==4.3.2
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.3
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.59.0
gunicorn==20.1.0
HeapDict==1.0.1
huggingface-hub==0.17.3
idna==3.4
ijson==3.2.3
importlib-metadata==6.8.0
importlib-resources==6.1.0
inscriptis==2.3.2
ipykernel==6.25.2
ipython==8.12.3
ir-datasets==0.5.5
ir-measures==0.3.3
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.5
lxml==4.9.3
lz4==4.3.2
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
matchpy==0.5.5
matplotlib==3.7.3
matplotlib-inline==0.1.6
mlflow==2.5.0
more-itertools==10.1.0
mpmath==1.3.0
msal==1.24.0
msal-extensions==1.0.0
multiset==2.1.1
nest-asyncio==1.5.8
networkx==3.1
nptyping==1.4.4
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
packaging==23.1
pandas==1.5.3
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pip==23.2.1
platformdirs==3.11.0
portalocker==2.8.2
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyautocorpus==0.1.12
pycparser==2.21
Pygments==2.16.1
PyHive==0.7.0
pyjnius==1.5.0
PyJWT==2.8.0
pyparsing==3.1.1
pyterrier_colbert @ git+https://github.com/terrierteam/pyterrier_colbert.git
python-arango==7.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-terrier==0.9.2
pytrec-eval-terrier==0.5.5
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
querystring-parser==1.2.4
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.10.1
sentencepiece==0.1.99
SetSimilaritySearch==1.0.1
setuptools==56.0.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==2.0.21
sqlparse==0.4.4
stack-data==0.6.3
statsmodels==0.14.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tika==2.6.0
tokenizers==0.13.3
torch==2.1.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.33.2
trec-car-tools==2.6
tree-sitter==0.20.1
triton==2.1.0
typing_extensions==4.8.0
typish==1.9.3
tzdata==2023.3
ujson==5.8.0
unlzw3==0.2.2
urllib3==1.26.16
warc3-wet==0.2.3
warc3-wet-clueweb09==0.2.5
wcwidth==0.2.8
websocket-client==1.6.4
Werkzeug==2.3.7
wget==3.2
wheel==0.41.2
wrapt==1.15.0
zipp==3.17.0
zlib-state==0.1.5


































